# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Happening API (Evently)
  version: 1.0.0
  description: |
    OpenAPI specification for the Happening/Evently backend. 
    This API demonstrates a scalable, production-grade architecture featuring:
    - **Redis Caching:** For all public GET routes to handle read-heavy traffic.
    - **RabbitMQ Queueing:** For the /bookings endpoint to handle write-heavy "flash sale" scenarios.
    - **PostgreSQL Transactions (FOR UPDATE):** To ensure data integrity and prevent overselling.
    - **Full Admin & Host Management:** Complete with approval flows and analytics.
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://your-deployed-url.com/api
    description: Production server

# ==========================================================
#  Reusable Components (Schemas & Security)
# ==========================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "User's JWT token, prefixed with 'Bearer '"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Invalid or expired token"
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        is_host:
          type: boolean
        host_status:
          type: string
          enum: [not_requested, pending, approved, rejected]
      example:
        id: "a1b2c3d4-..."
        email: "alice@example.com"
        name: "Alice"
        role: "user"
        is_host: true
        host_status: "approved"
...
    MyBooking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        ticket_count:
          type: integer
        total_amount:
          type: number
        event_id:
          type: string
          format: uuid
        event_title:
          type: string
        event_date:
          type: string
          format: date-time
        image_url:
          type: string
          format: uri
        category:
          type: string
        event_description:
          type: string
        payment_status:
          type: string
          enum: [pending, succeeded, refunded, refund_failed, no_payment]
...
    AuthRegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
      example:
        email: "alice@example.com"
        password: "pass123"
        name: "Alice"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      example:
        email: "alice@example.com"
        password: "pass123"
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      example:
        message: "User registered successfully"
        user:
          id: "a1b2c3d4-..."
          email: "alice@example.com"
          name: "Alice"
          role: "user"
        token: "eyJhbGci...token"
    BookingRequest:
      type: object
      required: [event_id, ticket_count]
      properties:
        event_id:
          type: string
          format: uuid
        ticket_count:
          type: integer
          minimum: 1
      example:
        event_id: "e1f2a3b4-..."
        ticket_count: 2
    BookingResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      example:
        success: true
        message: "Booking request received! We are processing your tickets."
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request (e.g., user exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      summary: Login and get JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized (Invalid credentials)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- Public Event Routes ---
...
  /bookings:
    post:
      summary: Enqueue a booking request (Async)
      description: This endpoint is asynchronous. It adds a job to the queue and returns 202 Accepted.
      tags: [Bookings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequest"
      responses:
        "202":
          description: Accepted (Booking is processing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "401":
          description: Unauthorized
        "503":
          description: Service unavailable (Queue is down)
  /bookings/my-bookings:
    get:
      summary: Get all bookings for the current user
      tags: [Bookings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: "#/components/schemas/MyBooking"
        "401":
          description: Unauthorized
  /bookings/{id}/cancel-pending:
    post:
      summary: Cancel a *pending* (unpaid) booking
      tags: [Bookings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Pending booking not found
          
  # --- Protected Cancellation Routes ---
...